requirements.txt
streamlit
numpy
opencv-python-headless
Pillow
openai
